+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @workflow/vite
 
+## 5.0.0-beta.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@5.0.0-beta.22
+
 ## 5.0.0-beta.21
 
 ### Patch Changes